predictserver/requirements.txt
matplotlib==3.9.2
numpy==1.26.4
Pillow==11.0.0
requests==2.32.3
scikit-learn==1.5.2
tensorflow==2.18.0

# Django==5.1.3
Flask==3.0.3

# jupyter==1.1.0
# seaborn==0.13.2
